Add a --date-limit option
The purpose of this is to limit the publication date of package versions
to be considered. By disallowing versions after a date, you can hope to
re-install the packages that would have been installed on that date,
which helps reproducing installs.

This is similar to the checkpoint() feature for R.